Split model into different py file
