+++ hr_expense_product_policy/models/__init__.py
@@ -2,5 +2,7 @@
 # © 2016 OpenSynergy Indonesia
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
+from . import hr_job
+from . import hr_department
 from . import hr_employee
 from . import hr_expense
